private-agent-router/requirements.txt
fastapi>=0.110
uvicorn>=0.29
pynacl>=1.5
httpx>=0.27
web3>=6.15
aiosqlite>=0.20
